requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    make pin-requirements
#
accept-types==0.4.1
    # via karapace (/karapace/pyproject.toml)
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.10.11
    # via karapace (/karapace/pyproject.toml)
aiokafka==0.10.0
    # via karapace (/karapace/pyproject.toml)
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-lru==2.0.4
    # via karapace (/karapace/pyproject.toml)
async-timeout==5.0.1
    # via
    #   aiohttp
    #   aiokafka
attrs==25.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
avro @ https://github.com/aiven/avro/archive/5a82d57f2a650fd87c819a30e433f1abb2c76ca2.tar.gz#subdirectory=lang/py
    # via karapace (/karapace/pyproject.toml)
cachetools==5.3.3
    # via karapace (/karapace/pyproject.toml)
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
confluent-kafka==2.4.0
    # via karapace (/karapace/pyproject.toml)
cramjam==2.9.1
    # via python-snappy
dependency-injector==4.43.0
    # via karapace (/karapace/pyproject.toml)
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi[standard]==0.115.5
    # via karapace (/karapace/pyproject.toml)
fastapi-cli[standard]==0.0.7
    # via fastapi
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
googleapis-common-protos==1.66.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.70.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.4.0
    # via opentelemetry-api
isodate==0.7.2
    # via karapace (/karapace/pyproject.toml)
jinja2==3.1.5
    # via fastapi
jsonschema==4.23.0
    # via karapace (/karapace/pyproject.toml)
jsonschema-specifications==2024.10.1
    # via jsonschema
lz4==4.4.3
    # via karapace (/karapace/pyproject.toml)
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via karapace (/karapace/pyproject.toml)
opentelemetry-api==1.27.0
    # via
    #   karapace (/karapace/pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.27.0
    # via karapace (/karapace/pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.27.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.27.0
    # via opentelemetry-exporter-otlp
opentelemetry-proto==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.27.0
    # via
    #   karapace (/karapace/pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.48b0
    # via opentelemetry-sdk
packaging==24.2
    # via aiokafka
prometheus-client==0.20.0
    # via karapace (/karapace/pyproject.toml)
protobuf==3.20.3
    # via
    #   googleapis-common-protos
    #   karapace (/karapace/pyproject.toml)
    #   opentelemetry-proto
pydantic==2.10.2
    # via
    #   fastapi
    #   karapace (/karapace/pyproject.toml)
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-settings==2.6.1
    # via karapace (/karapace/pyproject.toml)
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via karapace (/karapace/pyproject.toml)
python-dateutil==2.9.0.post0
    # via karapace (/karapace/pyproject.toml)
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
python-snappy==0.7.3
    # via karapace (/karapace/pyproject.toml)
pyyaml==6.0.2
    # via uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via opentelemetry-exporter-otlp-proto-http
rich==13.7.1
    # via
    #   karapace (/karapace/pyproject.toml)
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   dependency-injector
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
tenacity==9.0.0
    # via karapace (/karapace/pyproject.toml)
typer==0.15.1
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   async-lru
    #   fastapi
    #   karapace (/karapace/pyproject.toml)
    #   multidict
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   referencing
    #   rich-toolkit
    #   typer
    #   uvicorn
ujson==5.10.0
    # via karapace (/karapace/pyproject.toml)
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via
    #   karapace (/karapace/pyproject.toml)
    #   uvicorn
websockets==14.2
    # via uvicorn
wrapt==1.17.2
    # via deprecated
xxhash==3.5.0
    # via karapace (/karapace/pyproject.toml)
yarl==1.12.1
    # via
    #   aiohttp
    #   karapace (/karapace/pyproject.toml)
zipp==3.21.0
    # via importlib-metadata
zstandard==0.23.0
    # via karapace (/karapace/pyproject.toml)
